chore: Add support for selecting video quality in download API
The code changes modify the download API to include a new query parameter `quality` which allows the user to specify the desired video quality. The API now checks the value of the `quality` parameter and chooses the appropriate format for the video download. The `quality` parameter is also added to the frontend, allowing the user to select the desired quality from a dropdown menu.

Note: The commit message has been generated based on the provided code changes and recent commits.