fix: Update ffmpeg path to use static version
The code changes in the download API update the ffmpeg path to use the static version provided by the "ffmpeg-static" dependency. This ensures that the correct version of ffmpeg is used for processing videos. The "ffmpeg-static" package is added as a new dependency in the "package.json" file.

Note: The commit message has been generated based on the provided code changes and recent commits.